print: Add a generic wrapper for objects which supply Print and Printf
I want this so I can convert a *logrus.Logger (that is not
logrus.StandardLogger() [1]) to a log.Logger interface.

You can almost do the same thing with the fmt implementation.
Unfortunately, imported packages do not seem to be first-class objects
in Go, with:

  import (
    "fmt"

    "github.com/go-log/log/print"
  )

  ...

  print.New(fmt)

raising:

  use of package fmt without selector

We could work around that if fmt supplied an analog to
logrus.StandardLogger(), but it doesn't as of Go 1.10.3.

[1]: https://godoc.org/github.com/sirupsen/logrus#pkg-constants