nest: Add a sub-package for creating sub-loggers
Some libraries may want to add structure to child logs.  For example,
say your library takes a Logger and it turns around and passes that
Logger down into some child calls that are processing a request for
https://example.com/some/resource.  Those child calls don't care about
the original URI, but you want that URI to show up in anything logged
by the children for context.  With this package, you can set that up
with:

  func middleman(logger log.Logger) error {
    uri := "https://example.com/some/resource"
    resp, err := http.Get(uri)
    if err != nil {
      return err
    }
    return child(resp, nest.New(" ", uri, " "))
  }

The buffering spaces are a bit awkward, but several loggers follow
fmt.Sprint, which only adds spaces between operands when neither is a
string [1].  For example:

  $ cat test.go
  package main

  import (
    "flag"
    "log"

    "github.com/golang/glog"
    "github.com/sirupsen/logrus"
  )

  func main() {
    flag.Parse()
    log.Print("a", "b")
    glog.Info("c", "d")
    logrus.Info("e", "f")
    logrus.WithFields(map[string]interface{}{"key": "value", "count": 1}).Info("g", "h")
  }
  $ go run test.go -logtostderr
  2018/11/10 21:16:54 ab
  I1110 21:16:54.589271   13508 test.go:14] cd
  INFO[0000] ef
  INFO[0000] gh                                            count=1 key=value

So it's probably a good idea to add the separating spaces.  Structured
base loggers that preserve separate arguments should probably check
for and discard any arguments matching the single-space string.

As the unit tests show, New(...) can be used in the style of
logrus.WithFields to add structured information as well (although as
discussed above, the final output depends on how the base logger
handles Log(...) values).  You could certainly write a logrus wrapper
whose Log implementation that used WithFields on any arguments that
successfully cast to logrus.Fields.

[1]: https://golang.org/pkg/fmt/#Sprint
[2]: https://godoc.org/github.com/sirupsen/logrus#WithFields