log: Godocs for Logger methods
We want to give some guidance so that consumers use these consistently
and implementations know what they can expect to receive.  There's not
much flexibility for Logf, so I just punt to fmt.Printf there.  I
suggest that Log is going to be similar to fmt.Print as well, but
leave the door open for alternative implementations (see also
discussion in [1]).

[1]: https://github.com/go-log/log/issues/1