logrus: Drop this deprecated package
This package has been deprecated since the print package landed in
6fc7c2fc (print: Add a generic wrapper for objects which supply Print
and Printf, 2018-07-18, #6).  Removing the logrus package makes it
obvious that go-log/log consumers do not need to have a local copy of
logrus (which is kind of the point of go-log/log ;), even if they use
dependency tooling that does not support pruning unused subpackages.